# docs/swagger/swagger.yaml
definitions:
...
  Error:
    properties:
      code:
        example: INVALID_ARGUMENT
        type: string
      message:
        example: invalid request body
        type: string
      meta:
        additionalProperties:
          type: string
        example:
          error: invalid_length
          field: username
        type: object
    type: object
...
  GetMyProfileOutput:
    properties:
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      headline:
        type: string
      last_name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  OrganizationOutput:
    properties:
      active:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  SaveMyProfileInput:
    properties:
      bio:
        maxLength: 1000
        type: string
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      headline:
        maxLength: 100
        type: string
      last_name:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - email
    - first_name
    - last_name
    type: object
...
  UpdateEmployeeInput:
    properties:
      employment_type:
        enum:
        - full_time
        - part_time
        - contract
        - intern
        - temporary
        - consultant
        type: string
      job_title:
        maxLength: 100
        minLength: 2
        type: string
      membership_status:
        enum:
        - active
        - on_leave
        - suspended
        - resigned
        - terminated
        type: string
      role_id:
        type: string
    required:
    - employment_type
    - job_title
    - membership_status
    - role_id
    type: object
  UpdateOrganizationInput:
    properties:
      category:
        enum:
        - technology
        - finance
        - retail
        - manufacturing
        - other
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - category
    - name
    type: object
...
info:
  contact:
    email: support@trueflow.ir
    name: API Support
    url: https://trueflow.ir/support
  description: TrueFlow is a modern human resource management (HRM) platform designed
    to integrate organizational processes, manage teams, control access, and increase
    organizational productivity.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: TrueFlow Api Document
paths:
...
  /organization/{organization_id}:
    get:
      description: Returns the business details of an organization.
      parameters:
      - description: Organization ID
        in: path
        name: organization_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OrganizationOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: Get organization
      tags:
      - Organization
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: organization_id
        required: true
        type: string
      - description: Update organization request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateOrganizationInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: Update organization
      tags:
      - Organization
...
  /organization/{organization_id}/employee/{employee_id}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: organization_id
        required: true
        type: string
      - description: Employee ID
        in: path
        name: employee_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: Remove employee
      tags:
      - Employee
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: organization_id
        required: true
        type: string
      - description: Employee ID
        in: path
        name: employee_id
        required: true
        type: string
      - description: Update employee request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateEmployeeInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: Update employee
      tags:
      - Employee
...
  /user/me/profile:
    get:
      description: Returns the authenticated user's profile information.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetMyProfileOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: Get current user's profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Creates or updates the authenticated user's profile.
      parameters:
      - description: Profile
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SaveMyProfileInput'
      responses:
        "204":
          description: Profile updated successfully
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: Create or update current user's profile
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Enter the JWT token with the `Bearer ` prefix
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"